move all functions located in <pthread_sleep.c> to this file
