+++ components/libc/compilers/common/unistd.c
@@ -5,10 +5,13 @@
  *
  * Change Logs:
  * Date           Author       Notes
- * 2020-09-01     Meco Man     First Version
+ * 2020-09-01     Meco Man     first Version
+ * 2021-02-12     Meco Man     move all functions located in <pthread_sleep.c> to this file
  */
 
 #include <unistd.h>
+#include <rtthread.h>
+#include <rthw.h>
 
 #ifdef RT_USING_POSIX_TERMIOS
 #include "termios.h"
@@ -23,4 +26,22 @@ int isatty(int fd)
 char *ttyname(int fd)
 {
     return "/dev/tty0"; /*TODO: need to add more specific*/
+}
+
+unsigned int sleep(unsigned int seconds)
+{
+    rt_tick_t delta_tick;
+
+    delta_tick = rt_tick_get();
+    rt_thread_delay(seconds * RT_TICK_PER_SECOND);
+    delta_tick = rt_tick_get() - delta_tick;
+
+    return seconds - delta_tick/RT_TICK_PER_SECOND;
+}
+
+int usleep(useconds_t usec)
+{
+    rt_thread_mdelay(usec / 1000u);
+    rt_hw_us_delay(usec % 1000u);
+    return 0;
 }
